+++ packages/faceted-search/src/dictionary/badge.dictionary.js
@@ -1,10 +1,10 @@
-import { BadgeText } from '../components/Badges/BadgeText/BadgeText.component';
 import { BadgeCheckboxes } from '../components/Badges/BadgeCheckboxes/BadgeCheckboxes.component';
-import { BadgeNumber } from '../components/Badges/BadgeNumber/BadgeNumber.component';
-import { BadgeSlider } from '../components/Badges/BadgeSlider/BadgeSlider.component';
-import { BadgeTags } from '../components/Badges/BadgeTags/BadgeTags.component';
 import { BadgeDate } from '../components/Badges/BadgeDate/BadgeDate.component';
 import { BadgeMenu } from '../components/Badges/BadgeMenu/BadgeMenu.component';
+import { BadgeNumber } from '../components/Badges/BadgeNumber/BadgeNumber.component';
+import { BadgePeriod } from '../components/Badges/BadgePeriod/BadgePeriod.component';
+import { BadgeSlider } from '../components/Badges/BadgeSlider/BadgeSlider.component';
+import { BadgeText } from '../components/Badges/BadgeText/BadgeText.component';
 
 const standardBadgeTypeNames = {
 	text: 'text',
@@ -14,6 +14,7 @@ const standardBadgeTypeNames = {
 	tags: 'tags',
 	date: 'date',
 	menu: 'menu',
+	period: 'period',
 };
 
 const standardBadges = {
@@ -21,9 +22,10 @@ const standardBadges = {
 	[standardBadgeTypeNames.checkbox]: BadgeCheckboxes,
 	[standardBadgeTypeNames.number]: BadgeNumber,
 	[standardBadgeTypeNames.slider]: BadgeSlider,
-	[standardBadgeTypeNames.tags]: BadgeTags,
+	[standardBadgeTypeNames.tags]: BadgeCheckboxes,
 	[standardBadgeTypeNames.date]: BadgeDate,
 	[standardBadgeTypeNames.menu]: BadgeMenu,
+	[standardBadgeTypeNames.period]: BadgePeriod,
 };
 
 export const filterBadgeDefinitionsWithDictionary = (badgesDictionary, badgeDefinition) => {